Revert "updating analytics code"
