+++ .vscode/settings.json
@@ -11,5 +11,5 @@
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.formatOnSave": true,
   "prettier.configPath": ".prettierrc",
-  "prettier.prettierPath": "./node_modules/prettier"
+  "prettier.prettierPath": "node_modules/prettier"
 }
